toolSHare/requirements.txt
appnope==0.1.3
asgiref==3.6.0
astroid==2.15.5
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
Django==3.2.19
django-bootstrap4==2.2.0
djangorestframework==3.11.1
docopt==0.6.2
duckdb==0.8.0
executing==1.2.0
factory-boy==3.0.1
Faker==18.11.2
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.24.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==3.5.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
Pygments==2.15.1
pylance==0.4.14
pylint==2.17.4
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0
